squash -Wsign-conversion warnings in pic_string
This is the equivalent of 5a5aa9c606db77f444d744a11de61eb75de2b6ea and
be2de1fec0e5279a54a16e66cd96cdb620029a30 which we need to apply for a third time
because this function has been copied and pasted. Related to #2051.

+++ plugin/core/gvrender_core_pic.c
@@ -149,8 +149,8 @@ static void picptarray(GVJ_t *job, pointf * A, int n, int close)
 static char *pic_string(char *s)
 {
     static char *buf = NULL;
-    static int bufsize = 0;
-    int pos = 0;
+    static size_t bufsize = 0;
+    size_t pos = 0;
     char *p;
     unsigned char c;
